+++ common/src/main/java/dev/terminalmc/effecttimerplus/gui/screen/ConfigScreenProvider.java
@@ -23,7 +23,8 @@ import net.minecraft.client.gui.components.MultiLineTextWidget;
 import net.minecraft.client.gui.screens.ConfirmLinkScreen;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.CommonComponents;
-import net.minecraft.util.Util;
+
+import java.net.URI;
 
 import static dev.terminalmc.effecttimerplus.util.Localization.localized;
 
@@ -39,8 +40,8 @@ public class ConfigScreenProvider {
 
     public static Screen getConfigScreen(Screen parent) {
         try {
-            return YaclScreenProvider.getConfigScreen(parent);
-//            return new DisabledScreen(parent);
+//            return YaclScreenProvider.getConfigScreen(parent);
+            return new DisabledScreen(parent);
         } catch (NoClassDefFoundError ignored) {
             return new BackupScreen(parent, "installYacl", "https://modrinth.com/project/1eAoo2KR");
         }
@@ -73,13 +74,7 @@ public class ConfigScreenProvider {
 
             Button openLinkButton = Button.builder(
                             localized("message", "viewModrinth"),
-                            (button) -> Minecraft.getInstance().gui.setScreen(new ConfirmLinkScreen(
-                                    (open) -> {
-                                        if (open)
-                                            Util.getPlatform().openUri(modUrl);
-                                        onClose();
-                                    }, modUrl, true
-                            ))
+                            ConfirmLinkScreen.confirmLink(this.parent, URI.create(modUrl), true)
                     )
                     .pos(width / 2 - 120, height / 2)
                     .size(115, 20)