+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.104.0-beta.1
+
+- Updated to mc26.3-rc-2
+- Temporarily disabled config screen
+
 ## 3.103.0
 
 - Updated to mc26.2